+++ client/src/comps/OrdersTable.jsx
@@ -75,7 +75,12 @@ const OrdersTable = () => {
   };
 
   const deleteOrder = async (oid) => {
+    const confimation = confirm("Do you really want to delete?");
+
+    if (!confimation) return;
+
     setLoading(true);
+
     try {
       const res = await axios.delete(
         `http://localhost:8000/api/v1/delete-order/${oid}`,
